+++ src/core/cue-handler.ts
@@ -1,13 +1,10 @@
-import {Overlaps} from './components';
 import {GameState} from './game-state';
 import {CuePayload, CueTypes, Setting} from './setting';
 
 export interface CueHandler<Stg extends Setting, Type extends CueTypes<Stg>> {
   generateCuesAtUpdate(
     state: GameState<Stg>,
-    args: {
-      overlaps: Overlaps;
-    }
+    args: {}
   ): CuePayload<Stg, Type>[];
 
   applyCue(